requirements.txt
littellm
pandas
pyodbc
streamlit
python-dotenv